refactor: consolidate auth middleware logic into middleware.dart and remove dedicated file
